Smoothness 2.0
Removed press down for strong attacks, restructured the order of the combos (easier to get combos) changed animation speed and flow, smoother now, added movement swipe edits to movement (will implement properly with all sprites). Adjusted jump so that there is a build up to the jump, and a cool down. Adjusted healthbar so that every knockback also comes with lowered health (there was a discrepancy between the max health (400) and the healthbar display (300)). removed background music temporarily just to focus on tweaking gameplay. Adjusted gravity and jump speed, adds more weight to the jump. Adjusted script so that animation end event is not used, have begun to implement damage hitbox, but will continue to adjust. Removed acceleration feature in order to maintain balance between walking speed and walking animation, adjusted borders of world so that players may not walk over background objects.